fix float/int typing, verify against live api responses

+++ src/com/amazon/creators/model/BrowseNode.php
@@ -49,7 +49,7 @@ class BrowseNode implements ModelInterface, ArrayAccess, \JsonSerializable
         'displayName' => 'string',
         'id' => 'string',
         'isRoot' => 'bool',
-        'salesRank' => 'float'
+        'salesRank' => 'int'
     ];
 
     /**
@@ -482,9 +482,9 @@ class BrowseNode implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Gets salesRank
      *
-     * @return float|null
+     * @return int|null
      */
-    public function getSalesRank(): float|null
+    public function getSalesRank(): int|null
     {
         return $this->container['salesRank'];
     }
@@ -492,11 +492,11 @@ class BrowseNode implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Sets salesRank
      *
-     * @param float|null $salesRank salesRank
+     * @param int|null $salesRank salesRank
      *
      * @return self
      */
-    public function setSalesRank(?float $salesRank): self
+    public function setSalesRank(?int $salesRank): self
     {
         if (is_null($salesRank)) {
             throw new \InvalidArgumentException('non-nullable salesRank cannot be null');
