+++ src/com/amazon/creators/model/Feed.php
@@ -42,7 +42,7 @@ class Feed implements ModelInterface, ArrayAccess, \JsonSerializable
       */
     protected static $openAPITypes = [
         'feedName' => 'string',
-        'size' => 'float',
+        'size' => 'int',
         'md5' => 'string',
         'lastUpdated' => 'string'
     ];
@@ -336,9 +336,9 @@ class Feed implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Gets size
      *
-     * @return float
+     * @return int
      */
-    public function getSize(): float
+    public function getSize(): int
     {
         return $this->container['size'];
     }
@@ -346,11 +346,11 @@ class Feed implements ModelInterface, ArrayAccess, \JsonSerializable
     /**
      * Sets size
      *
-     * @param float $size size
+     * @param int $size size
      *
      * @return self
      */
-    public function setSize(float $size): self
+    public function setSize(int $size): self
     {
         if (is_null($size)) {
             throw new \InvalidArgumentException('non-nullable size cannot be null');
